refactor removing backend split in code org. this was vestigial from old dataclass support that is not present anymore.
