+++ spock/addons/tune/ax.py
@@ -3,6 +3,6 @@
 # Copyright FMR LLC <opensource@fidelity.com>
 # SPDX-License-Identifier: Apache-2.0
 
-"""Handles building an adapter to the hyperopt backend"""
+"""Handles the ax backend"""
