+++ spock/args.py
@@ -6,4 +6,4 @@
 """Handles import aliases to allow backwards compat with backends"""
 
 # from spock.backend.dataclass.args import *
-from spock.backend.attr.typed import SavePath
+from spock.backend.typed import SavePath
