--- spock/backend/attr/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright FMR LLC <opensource@fidelity.com>
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-Spock is a framework that helps manage complex parameter configurations for Python applications
-
-Please refer to the documentation provided in the README.md
-"""
-
-__all__ = ["builder", "config", "payload", "saver", "typed"]
