+++ spock/backend/utils.py
