+++ spock/utils.py
@@ -57,7 +57,7 @@ def _is_spock_instance(__obj: object):
         bool
 
     """
-    return (__obj.__module__ == 'spock.backend.attr.config') and attr.has(__obj)
+    return (__obj.__module__ == 'spock.backend.config') and attr.has(__obj)
 
 
 def make_argument(arg_name, arg_type, parser):
